third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/icon-blocked-expected.txt

CONSOLE ERROR: line 15: Refused to load the image 'http://localhost/foo?q=from_icon' because it violates the following Content Security Policy directive: "img-src 'none'".

Use callbacks to show that favicons are not loaded in violation of CSP when link tags are dynamically added to the page.
